# survey-maker-backend/Application/Dtos/AuthDtos.cs
using System.ComponentModel.DataAnnotations;

namespace SurveyMaker.Api.Application.Dtos
{
    public class LoginRequestDto
    {
        [Required]
        [EmailAddress]
        public string Email { get; set; } = string.Empty;

        [Required]
        public string Password { get; set; } = string.Empty;
    }
}